Moved sockaddr function from packing to net.
