+++ pwnlib/shellcraft/templates/amd64/linux/connect.asm
@@ -1,5 +1,5 @@
 <% from pwnlib.shellcraft import amd64 %>
-<% from pwnlib.util import packing %>
+<% from pwnlib.util.net import sockaddr %>
 
 <%page args="host, port, network = 'ipv4'"/>
 <%docstring>
@@ -8,7 +8,7 @@
     Leaves the connected socket in rbp.
 </%docstring>
 <%
-    sockaddr, address_family = packing.sockaddr(host, port, network)
+    sockaddr, address_family = sockaddr(host, port, network)
 %>\
     /* open new socket */
     ${amd64.linux.syscall('SYS_socket', address_family, 'SOCK_STREAM', 0)}
